requirements.txt
pandas
requests
Biopython 
freesasa